+++ wezterm-gui/src/termwindow/mouseevent.rs
@@ -57,7 +57,7 @@ impl super::TermWindow {
 
         self.current_mouse_event.replace(event.clone());
 
-        let first_line_offset = if self.show_tab_bar && !self.config.tab_bar_at_bottom {
+        let first_line_offset = if self.show_tab_bar && !self.config.tab_bar_at_bottom() {
             self.tab_bar_pixel_height().unwrap_or(0.) as isize
         } else {
             0
@@ -235,7 +235,7 @@ impl super::TermWindow {
             current_viewport,
             &self.dimensions,
             self.tab_bar_pixel_height().unwrap_or(0.),
-            self.config.tab_bar_at_bottom,
+            self.config.tab_bar_at_bottom(),
         );
         self.set_viewport(pane.pane_id(), Some(row), dims);
         context.invalidate();